WTA #40: Removed old gradient colors.
