+++ core/ui/src/main/java/com/jacob/wakatimeapp/core/ui/theme/WakaTimeAppTheme.kt
@@ -9,26 +9,20 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.ReadOnlyComposable
 import com.jacob.wakatimeapp.core.ui.theme.assets.Assets
 import com.jacob.wakatimeapp.core.ui.theme.assets.LocalAssets
-import com.jacob.wakatimeapp.core.ui.theme.colors.Gradients
 import com.jacob.wakatimeapp.core.ui.theme.colors.LocalGradients
-import com.jacob.wakatimeapp.core.ui.theme.colors.LocalNewGradients
 import com.jacob.wakatimeapp.core.ui.theme.colors.dark.DarkColors
 import com.jacob.wakatimeapp.core.ui.theme.colors.dark.DarkGradients
-import com.jacob.wakatimeapp.core.ui.theme.colors.dark.DarkNewGradients
 import com.jacob.wakatimeapp.core.ui.theme.colors.light.LightColors
 import com.jacob.wakatimeapp.core.ui.theme.colors.light.LightGradients
-import com.jacob.wakatimeapp.core.ui.theme.colors.light.LightNewGradients
 
 @Composable
 fun WakaTimeAppTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable () -> Unit) {
     val colors = if (darkTheme) DarkColors else LightColors
     val gradients = if (darkTheme) DarkGradients else LightGradients
-    val newGradients = if (darkTheme) DarkNewGradients else LightNewGradients
 
     CompositionLocalProvider(
         LocalSpacing provides Spacing,
         LocalGradients provides gradients,
-        LocalNewGradients provides newGradients,
         LocalAssets provides Assets,
     ) {
         MaterialTheme(
@@ -40,17 +34,17 @@ fun WakaTimeAppTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Compo
     }
 }
 
-val MaterialTheme.spacing: Spacing
+val MaterialTheme.spacing
     @Composable
     @ReadOnlyComposable
     get() = LocalSpacing.current
 
-val MaterialTheme.gradients: Gradients
+val MaterialTheme.gradients
     @Composable
     @ReadOnlyComposable
     get() = LocalGradients.current
 
-val MaterialTheme.assets: Assets
+val MaterialTheme.assets
     @Composable
     @ReadOnlyComposable
     get() = LocalAssets.current
