+++ core/ui/src/main/java/com/jacob/wakatimeapp/core/ui/theme/colors/dark/Gradients.kt
@@ -6,119 +6,149 @@ import androidx.compose.ui.graphics.Color
 import com.jacob.wakatimeapp.core.ui.theme.colors.Gradient
 import com.jacob.wakatimeapp.core.ui.theme.colors.Gradients
 
-private val BLUE = Color(0xFF3F67E0)
-private val CYAN = Color(0xFF65D1E2)
-private val DARK_PINK = Color(0xFFCC1FBB)
-private val GREEN = Color(0xFF18A963)
-private val PURPLE = Color(0xFF7B61FF)
-private val RED = Color(0xFFEB4F4F)
-private val PINK = Color(0xFFDC52B5)
-private val DARK_PURPLE = Color(0xFF921FC9)
-private val CYAN_ALT = Color(0xFF5EC2D2)
-private val ORANGE = Color(0xFFFF9900)
-private val YELLOW = Color(0xFFE8C754)
+private val flareStart = Color(0xFFFFB4A7)
+private val onFlareStart = Color(0xFF670400)
+private val flareStartContainer = Color(0xFF910900)
+private val onFlareStartContainer = Color(0xFFFFDAD4)
 
-private val orange = Color(0xFFFFB86F)
-private val onOrange = Color(0xFF4A2800)
-private val orangeContainer = Color(0xFF693C00)
-private val onOrangeContainer = Color(0xFFFFDCBD)
+private val flareEnd = Color(0xFFFFBA33)
+private val onFlareEnd = Color(0xFF422C00)
+private val flareEndContainer = Color(0xFF5F4100)
+private val onFlareEndContainer = Color(0xFFFFDEAB)
 
-private val yellow = Color(0xFFE6C449)
-private val onYellow = Color(0xFF3B2F00)
-private val yellowContainer = Color(0xFF564500)
-private val onYellowContainer = Color(0xFFFFE07D)
+private val shifterStart = Color(0xFFFFADE0)
+private val onShifterStart = Color(0xFF60004D)
+private val shifterStartContainer = Color(0xFF801868)
+private val onShifterStartContainer = Color(0xFFFFD8ED)
 
-private val blue = Color(0xFFB6C4FF)
-private val onBlue = Color(0xFF00287D)
-private val blueContainer = Color(0xFF003BB0)
-private val onBlueContainer = Color(0xFFDCE1FF)
+private val shifterEnd = Color(0xFFFFB2BA)
+private val onShifterEnd = Color(0xFF670020)
+private val shifterEndContainer = Color(0xFF910030)
+private val onShifterEndContainer = Color(0xFFFFD9DC)
 
-private val cyan = Color(0xFF4FD8EC)
-private val onCyan = Color(0xFF00363D)
-private val cyanContainer = Color(0xFF004F58)
-private val onCyanContainer = Color(0xFF99F0FF)
+private val quepalStart = Color(0xFF51DBCD)
+private val onQuepalStart = Color(0xFF003733)
+private val quepalStartContainer = Color(0xFF00504A)
+private val onQuepalStartContainer = Color(0xFF72F8E9)
 
-private val darkPink = Color(0xFFFFACEB)
-private val onDarkPink = Color(0xFF5D0055)
-private val darkPinkContainer = Color(0xFF840079)
-private val onDarkPinkContainer = Color(0xFFFFD7F1)
+private val quepalEnd = Color(0xFF23E373)
+private val onQuepalEnd = Color(0xFF003917)
+private val quepalEndContainer = Color(0xFF005224)
+private val onQuepalEndContainer = Color(0xFF64FF92)
 
-private val green = Color(0xFF5EDE92)
-private val onGreen = Color(0xFF00391D)
-private val greenContainer = Color(0xFF00522C)
-private val onGreenContainer = Color(0xFF7CFBAC)
+private val tealLoveStart = Color(0xFF87D988)
+private val onTealLoveStart = Color(0xFF00390E)
+private val tealLoveStartContainer = Color(0xFF005318)
+private val onTealLoveStartContainer = Color(0xFFA2F6A1)
 
-private val purple = Color(0xFFC9BFFF)
-private val onPurple = Color(0xFF2E009C)
-private val purpleContainer = Color(0xFF441CC8)
-private val onPurpleContainer = Color(0xFFE5DEFF)
+private val tealLoveEnd = Color(0xFF00E1A6)
+private val onTealLoveEnd = Color(0xFF003827)
+private val tealLoveEndContainer = Color(0xFF00513A)
+private val onTealLoveEndContainer = Color(0xFF3DFFC0)
 
-private val red = Color(0xFFFFB3AE)
-private val onRed = Color(0xFF68000D)
-private val redContainer = Color(0xFF910618)
-private val onRedContainer = Color(0xFFFFDAD7)
+private val facebookMessengerStart = Color(0xFF6DD2FF)
+private val onFacebookMessengerStart = Color(0xFF003547)
+private val facebookMessengerStartContainer = Color(0xFF004D65)
+private val onFacebookMessengerStartContainer = Color(0xFFBFE9FF)
 
-private val pink = Color(0xFFFFADE0)
-private val onPink = Color(0xFF60004C)
-private val pinkContainer = Color(0xFF87006C)
-private val onPinkContainer = Color(0xFFFFD8EC)
+private val facebookMessengerEnd = Color(0xFFAFC6FF)
+private val onFacebookMessengerEnd = Color(0xFF002D6D)
+private val facebookMessengerEndContainer = Color(0xFF00429A)
+private val onFacebookMessengerEndContainer = Color(0xFFD9E2FF)
 
-private val darkPurple = Color(0xFFE8B3FF)
-private val onDarkPurple = Color(0xFF500074)
-private val darkPurpleContainer = Color(0xFF7200A3)
-private val onDarkPurpleContainer = Color(0xFFF6D9FF)
+private val reefStart = Color(0xFF47D6FF)
+private val onReefStart = Color(0xFF003543)
+private val reefStartContainer = Color(0xFF004E60)
+private val onReefStartContainer = Color(0xFFB6EBFF)
 
-private val cyanAlt = Color(0xFF50D8EC)
-private val onCyanAlt = Color(0xFF00363D)
-private val cyanAltContainer = Color(0xFF004F58)
-private val onCyanAltContainer = Color(0xFF99F0FF)
+private val reefEnd = Color(0xFFA9C7FF)
+private val onReefEnd = Color(0xFF003063)
+private val reefEndContainer = Color(0xFF00468C)
+private val onReefEndContainer = Color(0xFFD6E3FF)
 
-val DarkGradients = Gradients(
-    primary = Gradient(
-        startColor = blue,
-        endColor = cyan,
-        onStartColor = onBlue,
-        onEndColor = onCyan
+private val aminStart = Color(0xFFDDB7FF)
+private val onAminStart = Color(0xFF4A0080)
+private val aminStartContainer = Color(0xFF6900B3)
+private val onAminStartContainer = Color(0xFFF0DBFF)
+
+private val aminEnd = Color(0xFFCABEFF)
+private val onAminEnd = Color(0xFF30009B)
+private val aminEndContainer = Color(0xFF4700D7)
+private val onAminEndContainer = Color(0xFFE6DEFF)
+
+private val neuromancerStart = Color(0xFFFFAEDD)
+private val onNeuromancerStart = Color(0xFF600049)
+private val neuromancerStartContainer = Color(0xFF880069)
+private val onNeuromancerStartContainer = Color(0xFFFFD8EB)
+
+private val neuromancerEnd = Color(0xFFFFB0CE)
+private val onNeuromancerEnd = Color(0xFF64003A)
+private val neuromancerEndContainer = Color(0xFF8C0054)
+private val onNeuromancerEndContainer = Color(0xFFFFD9E5)
+
+private val purpinkStart = Color(0xFFD5BAFF)
+private val onPurpinkStart = Color(0xFF42008A)
+private val purpinkStartContainer = Color(0xFF5E00C1)
+private val onPurpinkStartContainer = Color(0xFFECDCFF)
+
+private val purpinkEnd = Color(0xFFFBAAFF)
+private val onPurpinkEnd = Color(0xFF580064)
+private val purpinkEndContainer = Color(0xFF7C008D)
+private val onPurpinkEndContainer = Color(0xFFFFD6FC)
+
+internal val DarkGradients = Gradients(
+    flare = Gradient(
+        startColor = flareStart,
+        endColor = flareEnd,
+        onStartColor = onFlareStart,
+        onEndColor = onFlareEnd
     ),
-    secondary = Gradient(
-        startColor = Color(0xFF3F55E0),
-        endColor = Color(0xFF5EC2D2),
-        opacity = 85
+    shifter = Gradient(
+        startColor = shifterStart,
+        endColor = shifterEnd,
+        onStartColor = onShifterStart,
+        onEndColor = onShifterEnd
     ),
-    pinkCyanLight = Gradient(
-        startColor = pink,
-        endColor = cyan,
-        onStartColor = onPink,
-        onEndColor = onCyan
+    quepal = Gradient(
+        startColor = quepalStart,
+        endColor = quepalEnd,
+        onStartColor = onQuepalStart,
+        onEndColor = onQuepalEnd
     ),
-    purpleCyanDark = Gradient(
-        startColor = darkPurple,
-        endColor = cyanAlt,
-        onStartColor = onDarkPurple,
-        onEndColor = onCyanAlt
+    tealLove = Gradient(
+        startColor = tealLoveStart,
+        endColor = tealLoveEnd,
+        onStartColor = onTealLoveStart,
+        onEndColor = onTealLoveEnd
     ),
-    greenCyan = Gradient(
-        startColor = green,
-        endColor = cyan,
-        onStartColor = onGreen,
-        onEndColor = onCyan
+    facebookMessenger = Gradient(
+        startColor = facebookMessengerStart,
+        endColor = facebookMessengerEnd,
+        onStartColor = onFacebookMessengerStart,
+        onEndColor = onFacebookMessengerEnd
     ),
-    purpleCyan = Gradient(
-        startColor = purple,
-        endColor = cyan,
-        onStartColor = onPurple,
-        onEndColor = onCyan
+    reef = Gradient(
+        startColor = reefStart,
+        endColor = reefEnd,
+        onStartColor = onReefStart,
+        onEndColor = onReefEnd
     ),
-    redPurple = Gradient(
-        startColor = red,
-        endColor = purple,
-        onStartColor = onRed,
-        onEndColor = onPurple
+    amin = Gradient(
+        startColor = aminStart,
+        endColor = aminEnd,
+        onStartColor = onAminStart,
+        onEndColor = onAminEnd
     ),
-    orangeYellow = Gradient(
-        startColor = orange,
-        endColor = yellow,
-        onStartColor = onOrange,
-        onEndColor = onYellow
+    neuromancer = Gradient(
+        startColor = neuromancerStart,
+        endColor = neuromancerEnd,
+        onStartColor = onNeuromancerStart,
+        onEndColor = onNeuromancerEnd
+    ),
+    purpink = Gradient(
+        startColor = purpinkStart,
+        endColor = purpinkEnd,
+        onStartColor = onPurpinkStart,
+        onEndColor = onPurpinkEnd
     )
 )
