+++ core/ui/src/main/java/com/jacob/wakatimeapp/core/ui/theme/colors/light/Gradients.kt
@@ -6,103 +6,149 @@ import androidx.compose.ui.graphics.Color
 import com.jacob.wakatimeapp.core.ui.theme.colors.Gradient
 import com.jacob.wakatimeapp.core.ui.theme.colors.Gradients
 
-private val orange = Color(0xFF8A5100)
-private val onOrange = Color(0xFFFFFFFF)
-private val orangeContainer = Color(0xFFFFDCBD)
-private val onOrangeContainer = Color(0xFF2C1600)
+private val flareStart = Color(0xFFBE0F00)
+private val onFlareStart = Color(0xFFFFFFFF)
+private val flareStartContainer = Color(0xFFFFDAD4)
+private val onFlareStartContainer = Color(0xFF400200)
 
-private val yellow = Color(0xFF725C00)
-private val onYellow = Color(0xFFFFFFFF)
-private val yellowContainer = Color(0xFFFFE07D)
-private val onYellowContainer = Color(0xFF231B00)
+private val flareEnd = Color(0xFF7E5700)
+private val onFlareEnd = Color(0xFFFFFFFF)
+private val flareEndContainer = Color(0xFFFFDEAB)
+private val onFlareEndContainer = Color(0xFF281900)
 
-private val blue = Color(0xFF3463E4)
-private val onBlue = Color(0xFFFFFFFF)
-private val blueContainer = Color(0xFFDCE1FF)
-private val onBlueContainer = Color(0xFF00164F)
+private val shifterStart = Color(0xFF9D3481)
+private val onShifterStart = Color(0xFFFFFFFF)
+private val shifterStartContainer = Color(0xFFFFD8ED)
+private val onShifterStartContainer = Color(0xFF3B002E)
 
-private val cyan = Color(0xFF26C6DA)
-private val onCyan = Color(0xFFFFFFFF)
-private val cyanContainer = Color(0xFF99F0FF)
-private val onCyanContainer = Color(0xFF001F24)
+private val shifterEnd = Color(0xFFBD0041)
+private val onShifterEnd = Color(0xFFFFFFFF)
+private val shifterEndContainer = Color(0xFFFFD9DC)
+private val onShifterEndContainer = Color(0xFF400010)
 
-private val darkPink = Color(0xFFAC009E)
-private val onDarkPink = Color(0xFFFFFFFF)
-private val darkPinkContainer = Color(0xFFFFD7F1)
-private val onDarkPinkContainer = Color(0xFF390034)
+private val tealLoveEnd = Color(0xFF006C4E)
+private val onTealLoveEnd = Color(0xFFFFFFFF)
+private val tealLoveEndContainer = Color(0xFF3DFFC0)
+private val onTealLoveEndContainer = Color(0xFF002115)
 
-private val green = Color(0xFF00A75D)
-private val onGreen = Color(0xFFFFFFFF)
-private val greenContainer = Color(0xFF7CFBAC)
-private val onGreenContainer = Color(0xFF00210F)
+private val quepalStart = Color(0xFF006A62)
+private val onQuepalStart = Color(0xFFFFFFFF)
+private val quepalStartContainer = Color(0xFF72F8E9)
+private val onQuepalStartContainer = Color(0xFF00201D)
 
-private val purple = Color(0xFF6A4EE2)
-private val onPurple = Color(0xFFFFFFFF)
-private val purpleContainer = Color(0xFFE5DEFF)
-private val onPurpleContainer = Color(0xFF1A0063)
+private val quepalEnd = Color(0xFF006D32)
+private val onQuepalEnd = Color(0xFFFFFFFF)
+private val quepalEndContainer = Color(0xFF64FF92)
+private val onQuepalEndContainer = Color(0xFF00210B)
 
-private val red = Color(0xFFB4262C)
-private val onRed = Color(0xFFFFFFFF)
-private val RedContainer = Color(0xFFFFDAD7)
-private val onRedContainer = Color(0xFF410005)
+private val tealLoveStart = Color(0xFF196D29)
+private val onTealLoveStart = Color(0xFFFFFFFF)
+private val tealLoveStartContainer = Color(0xFFA2F6A1)
+private val onTealLoveStartContainer = Color(0xFF002105)
 
-private val pink = Color(0xFFA82288)
-private val onPink = Color(0xFFFFFFFF)
-private val pinkContainer = Color(0xFFFFD8EC)
-private val onPinkContainer = Color(0xFF3B002E)
+private val facebookMessengerStart = Color(0xFF006685)
+private val onFacebookMessengerStart = Color(0xFFFFFFFF)
+private val facebookMessengerStartContainer = Color(0xFFBFE9FF)
+private val onFacebookMessengerStartContainer = Color(0xFF001F2A)
 
-private val darkPurple = Color(0xFF921FC9)
-private val onDarkPurple = Color(0xFFFFFFFF)
-private val darkPurpleContainer = Color(0xFFF6D9FF)
-private val onDarkPurpleContainer = Color(0xFF310049)
+private val facebookMessengerEnd = Color(0xFF0058C9)
+private val onFacebookMessengerEnd = Color(0xFFFFFFFF)
+private val facebookMessengerEndContainer = Color(0xFFD9E2FF)
+private val onFacebookMessengerEndContainer = Color(0xFF001944)
 
-private val cyanAlt = Color(0xFF006874)
-private val onCyanAlt = Color(0xFFFFFFFF)
-private val cyanAltContainer = Color(0xFF99F0FF)
-private val onCyanAltContainer = Color(0xFF001F24)
+private val reefStart = Color(0xFF00677F)
+private val onReefStart = Color(0xFFFFFFFF)
+private val reefStartContainer = Color(0xFFB6EBFF)
+private val onReefStartContainer = Color(0xFF001F28)
 
-val LightGradients = Gradients(
-    primary = Gradient(
-        startColor = blue,
-        endColor = cyan,
-        onStartColor = onBlue,
-        onEndColor = onCyan
+private val reefEnd = Color(0xFF055DB6)
+private val onReefEnd = Color(0xFFFFFFFF)
+private val reefEndContainer = Color(0xFFD6E3FF)
+private val onReefEndContainer = Color(0xFF001B3D)
+
+private val aminStart = Color(0xFF8621DA)
+private val onAminStart = Color(0xFFFFFFFF)
+private val aminStartContainer = Color(0xFFF0DBFF)
+private val onAminStartContainer = Color(0xFF2C0050)
+
+private val aminEnd = Color(0xFF5F32F3)
+private val onAminEnd = Color(0xFFFFFFFF)
+private val aminEndContainer = Color(0xFFE6DEFF)
+private val onAminEndContainer = Color(0xFF1C0062)
+
+private val neuromancerStart = Color(0xFFB1018A)
+private val onNeuromancerStart = Color(0xFFFFFFFF)
+private val neuromancerStartContainer = Color(0xFFFFD8EB)
+private val onNeuromancerStartContainer = Color(0xFF3B002C)
+
+private val neuromancerEnd = Color(0xFFB3166E)
+private val onNeuromancerEnd = Color(0xFFFFFFFF)
+private val neuromancerEndContainer = Color(0xFFFFD9E5)
+private val onNeuromancerEndContainer = Color(0xFF3E0022)
+
+private val purpinkStart = Color(0xFF7D00FA)
+private val onPurpinkStart = Color(0xFFFFFFFF)
+private val purpinkStartContainer = Color(0xFFECDCFF)
+private val onPurpinkStartContainer = Color(0xFF270057)
+
+private val purpinkEnd = Color(0xFFA300B9)
+private val onPurpinkEnd = Color(0xFFFFFFFF)
+private val purpinkEndContainer = Color(0xFFFFD6FC)
+private val onPurpinkEndContainer = Color(0xFF36003E)
+
+internal val LightGradients = Gradients(
+    flare = Gradient(
+        startColor = flareStart,
+        endColor = flareEnd,
+        onStartColor = onFlareStart,
+        onEndColor = onFlareEnd
     ),
-    secondary = Gradient(Color(0xFF3F55E0), Color(0xFF5EC2D2), 85),
-    pinkCyanLight = Gradient(
-        startColor = darkPink,
-        endColor = cyan,
-        onStartColor = onDarkPink,
-        onEndColor = onCyan
+    shifter = Gradient(
+        startColor = shifterStart,
+        endColor = shifterEnd,
+        onStartColor = onShifterStart,
+        onEndColor = onShifterEnd
     ),
-    purpleCyanDark = Gradient(
-        startColor = darkPurple,
-        endColor = cyanAlt,
-        onStartColor = onDarkPurple,
-        onEndColor = onCyanAlt
+    quepal = Gradient(
+        startColor = quepalStart,
+        endColor = quepalEnd,
+        onStartColor = onQuepalStart,
+        onEndColor = onQuepalEnd
     ),
-    greenCyan = Gradient(
-        startColor = green,
-        endColor = cyan,
-        onStartColor = onGreen,
-        onEndColor = onCyan
+    tealLove = Gradient(
+        startColor = tealLoveStart,
+        endColor = tealLoveEnd,
+        onStartColor = onTealLoveStart,
+        onEndColor = onTealLoveEnd
     ),
-    purpleCyan = Gradient(
-        startColor = purple,
-        endColor = cyan,
-        onStartColor = onPurple,
-        onEndColor = onCyan
+    facebookMessenger = Gradient(
+        startColor = facebookMessengerStart,
+        endColor = facebookMessengerEnd,
+        onStartColor = onFacebookMessengerStart,
+        onEndColor = onFacebookMessengerEnd
     ),
-    redPurple = Gradient(
-        startColor = red,
-        endColor = purple,
-        onStartColor = onRed,
-        onEndColor = onPurple
+    reef = Gradient(
+        startColor = reefStart,
+        endColor = reefEnd,
+        onStartColor = onReefStart,
+        onEndColor = onReefEnd
     ),
-    orangeYellow = Gradient(
-        startColor = orange,
-        endColor = yellow,
-        onStartColor = onOrange,
-        onEndColor = onYellow
+    amin = Gradient(
+        startColor = aminStart,
+        endColor = aminEnd,
+        onStartColor = onAminStart,
+        onEndColor = onAminEnd
+    ),
+    neuromancer = Gradient(
+        startColor = neuromancerStart,
+        endColor = neuromancerEnd,
+        onStartColor = onNeuromancerStart,
+        onEndColor = onNeuromancerEnd
+    ),
+    purpink = Gradient(
+        startColor = purpinkStart,
+        endColor = purpinkEnd,
+        onStartColor = onPurpinkStart,
+        onEndColor = onPurpinkEnd
     )
 )
